backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
